// src/theme/typography/types.ts
export interface TypographyVariantsOptions extends TypographyOptions {
  caps26: React.CSSProperties;
  caps24: React.CSSProperties;
  caps16: React.CSSProperties;
  caps15: React.CSSProperties;
  caps14: React.CSSProperties;
  caps12: React.CSSProperties;
  caps12Medium: React.CSSProperties;
  text24: React.CSSProperties;
  text22: React.CSSProperties;
  text20: React.CSSProperties;
  text18: React.CSSProperties;
  text16: React.CSSProperties;
  text15: React.CSSProperties;
  text14: React.CSSProperties;
  text12: React.CSSProperties;
  text10: React.CSSProperties;
  text14Medium: React.CSSProperties;
  text16Medium: React.CSSProperties;
  text18Medium: React.CSSProperties;
  text20Medium: React.CSSProperties;
  text24Medium: React.CSSProperties;
  text24Book: React.CSSProperties;
  text22Book: React.CSSProperties;
  text20Book: React.CSSProperties;
  text18Book: React.CSSProperties;
  text16Book: React.CSSProperties;
  text15Book: React.CSSProperties;
  text14Book: React.CSSProperties;
  text12Book: React.CSSProperties;
  text10Book: React.CSSProperties;
}
// ...
  interface TypographyOptions {
    caps26: React.CSSProperties;
    caps24: React.CSSProperties;
    caps16: React.CSSProperties;
    caps15: React.CSSProperties;
    caps14: React.CSSProperties;
    caps12: React.CSSProperties;
    caps12Medium: React.CSSProperties;
    text24: React.CSSProperties;
    text22: React.CSSProperties;
    text20: React.CSSProperties;
    text18: React.CSSProperties;
    text16: React.CSSProperties;
    text15: React.CSSProperties;
    text14: React.CSSProperties;
    text12: React.CSSProperties;
    text10: React.CSSProperties;
    text14Medium: React.CSSProperties;
    text16Medium: React.CSSProperties;
    text18Medium: React.CSSProperties;
    text20Medium: React.CSSProperties;
    text24Medium: React.CSSProperties;
    text24Book: React.CSSProperties;
    text22Book: React.CSSProperties;
    text20Book: React.CSSProperties;
    text18Book: React.CSSProperties;
    text16Book: React.CSSProperties;
    text15Book: React.CSSProperties;
    text14Book: React.CSSProperties;
    text12Book: React.CSSProperties;
    text10Book: React.CSSProperties;
  }